Move filesystem-related files to hitsumabushi_filesystem_linux.c
There is also a new FilesystemFilePath() function to override this file.
